+++ src/ClientInterface.php
@@ -7,6 +7,7 @@ namespace Aceproxies\ResellerApi;
 use Aceproxies\ResellerApi\Endpoint\BalanceInterface;
 use Aceproxies\ResellerApi\Endpoint\HealthInterface;
 use Aceproxies\ResellerApi\Endpoint\OrdersInterface;
+use Aceproxies\ResellerApi\Endpoint\ProductsInterface;
 
 interface ClientInterface
 {
@@ -15,6 +16,8 @@ interface ClientInterface
     public function balance(): BalanceInterface;
 
     public function orders(): OrdersInterface;
+
+    public function products(): ProductsInterface;
 
     public function getApiVersion(): string;
 }